+++ lib/commandline/cmdcommon/cmdcommon.ml
@@ -15,6 +15,8 @@
 (*                                                                                            *)
 (**********************************************************************************************)
 
+open Cmdliner
+
 let check_yomu_hidden () =
   match Libyomu.Init.check_yomu_hidden () with
   | Ok () ->
@@ -29,11 +31,80 @@ let check_yomu_initialiaze () =
   | Error _ ->
       raise @@ Libyomu.Error.(yomu_error @@ Yomu_Not_Initialized)
 
+let keep_unzipped_term =
+  let var_info = Cmd.Env.info Libyomu.App.KeyBindingConst.variable_keep_unzip in
+  Arg.(
+    value & flag
+    & info [ "keep-unzipped" ] ~env:var_info
+        ~doc:
+          "Indicate whether unzipped comics should be kept in memory. If set, \
+           unzipped comics won't be unzipped again if read again but cause a \
+           larger memory consumtion"
+  )
+
+let filter_dotfile ~path s =
+  match String.starts_with ~prefix:"." s with
+  | true ->
+      let () =
+        try Util.FileSys.rmrf (Filename.concat path s) () with _ -> ()
+      in
+      None
+  | false ->
+      Some s
+
 let password_prompt = "Enter the master password : "
+
+let ask_password_if_encrypted encrypted () =
+  match encrypted with
+  | false ->
+      None
+  | true ->
+      let () = check_yomu_hidden () in
+      let key =
+        Option.some
+        @@ Libyomu.Input.ask_password_encrypted ~prompt:password_prompt ()
+      in
+      key
+
+let make_variable_section variable content =
+  let variable = Printf.sprintf "$(b,%s)" variable in
+  `I (variable, content)
+
+let variable_description =
+  Printf.sprintf
+    "If this environment variable is present, the %s is mapped to the first \
+     letter of this environment variable"
+
+let var_next_page =
+  make_variable_section Libyomu.App.KeyBindingConst.key_variable_next_page
+  @@ variable_description "next page key"
+
+let var_previous_page =
+  make_variable_section Libyomu.App.KeyBindingConst.key_variable_previous_page
+  @@ variable_description "previous page key"
+
+let var_quit =
+  make_variable_section Libyomu.App.KeyBindingConst.key_variable_quit
+  @@ variable_description "quit key"
+
+let var_goto_page =
+  make_variable_section Libyomu.App.KeyBindingConst.key_variable_goto_page
+  @@ variable_description "goto page key"
+
+let var_goto_book =
+  make_variable_section Libyomu.App.KeyBindingConst.key_variable_goto_book
+  @@ variable_description "goto book key"
 
 let read_common_description =
   [
     `S "KEY BINDINGS";
+    `P "This section presents the default key bindings";
+    `Noblank;
+    `P
+      (Printf.sprintf
+         "see $(b,%s) section to see which environment variable maps which key"
+         Manpage.s_environment
+      );
     `I ("To go to the next page", "Press $(b,'l')");
     `I ("To go to the precious page", "Press $(b,'j')");
     `I ("To quit", "Press $(b,'q')");
@@ -43,4 +114,10 @@ let read_common_description =
     `P
       "If you had loaded multiple comics, you can do the same movement than \
        with the pages but with the book by replacing the 'g' by 'b'";
+    `S Manpage.s_environment;
+    var_next_page;
+    var_previous_page;
+    var_quit;
+    var_goto_page;
+    var_goto_book;
   ]