+++ lib/libyomu/drawing.ml
@@ -17,6 +17,8 @@
 
 open Cbindings
 
+type drawing_config = { keep_unzipped : bool }
+
 let debug_string content =
   Out_channel.with_open_text "debug" (fun oc -> Printf.fprintf oc "%s\n" content)
 
@@ -181,15 +183,15 @@ let read_choice () =
   let _ = Unix.read Unix.stdin bytes 0 len in
   let c = Bytes.get bytes 0 in
   match c with
-  | 'j' | 'J' ->
-      `Left
-  | 'l' | 'L' ->
+  | c when c = App.KeyBindingConst.val_previous_page ->
       `Right
-  | 'q' | 'Q' ->
+  | c when c = App.KeyBindingConst.val_next_page ->
+      `Left
+  | c when c = App.KeyBindingConst.val_quit ->
       `Quit
-  | 'b' ->
+  | c when c = App.KeyBindingConst.val_goto_book ->
       read_movement ~parser:parse_book_movement ()
-  | 'g' ->
+  | c when c = App.KeyBindingConst.val_goto_page ->
       read_movement ~parser:parser_page_movement ()
   | _ ->
       `Ignore
@@ -199,7 +201,7 @@ let read_page comic_name mode ignored index zipper =
   let (page : Comic.page option) = Zipper.top_left zipper in
   match page with
   | None ->
-      `Right
+      `Left
   | Some page ->
       let () =
         match ignored with
@@ -213,68 +215,80 @@ let read_page comic_name mode ignored index zipper =
       option
 
 let read_item mode (item : ('a, Comic.named_archive) Either.t) =
-  let (Comic.{ pages; name } as c) =
+  let ( let* ) = Option.bind in
+  let* (Comic.{ pages; name } as c) =
     match item with
     | Either.Left comic ->
-        comic
+        Some comic
     | Either.Right { name; archive_path } ->
-        let comic = Comic.CZip.comic_of_zip archive_path in
+        let* comic = Comic.CZip.comic_of_zip archive_path in
         let () = Gc.major () in
         let () = Gc.compact () in
-
-        { comic with name }
+        Some { comic with name }
   in
 
   let z_pages = Zipper.of_list pages in
   let res = Zipper.action 0 (read_page name mode) z_pages in
-  (c, res)
+  Some (c, res)
 
-let read_collection mode =
+let read_collection mode config =
   Zipper.action_alt (fun zipper ->
       let current_opt = Zipper.top_left zipper in
       match current_opt with
       | None ->
-          (zipper, `Right)
+          (zipper, `Left)
       | Some either_comic ->
-          let comic, res = read_item mode either_comic in
-          let zipper =
-            match either_comic with
-            | Either.Right _ ->
-                Zipper.replace_current (Either.left comic) zipper
-            | Either.Left _ ->
-                zipper
-          in
           let zipper, res =
-            match res with
-            | (`Left | `Quit | `Right) as e ->
-                (zipper, e)
-            | `GotoBook kind ->
-                let n, res =
-                  match kind.Zipper.offset with
-                  | n when n < 0 && not kind.absolute ->
-                      (1, `Right)
-                  | n when n > 0 && not kind.absolute ->
-                      (-1, `Left)
-                  | _ ->
-                      (-1, `Left)
+            match read_item mode either_comic with
+            | None ->
+                (Zipper.remove_current zipper, `NoAction)
+            | Some (comic, res) ->
+                let zipper =
+                  match either_comic with
+                  | Either.Right _ -> (
+                      match config.keep_unzipped with
+                      | true ->
+                          Zipper.replace_current (Either.left comic) zipper
+                      | false ->
+                          zipper
+                    )
+                  | Either.Left _ ->
+                      zipper
                 in
-                (* Need this offset [n] since the since [res] will also move the zipper by one so we remove one by the movement *)
-                let kind = { kind with offset = kind.offset + n } in
-                let zipper = Zipper.move kind zipper in
+                let zipper, res =
+                  match res with
+                  | (`Left | `Quit | `Right) as e ->
+                      (zipper, e)
+                  | `GotoBook kind ->
+                      let n, res =
+                        match kind.Zipper.offset with
+                        | n when n < 0 && not kind.absolute ->
+                            (1, `Right)
+                        | n when n > 0 && not kind.absolute ->
+                            (-1, `Left)
+                        | _ ->
+                            (0, `NoAction)
+                      in
+                      (* Need this offset [n] since the since [res] will also move the zipper by one so we remove one by the movement *)
+                      let kind = { kind with offset = kind.offset + n } in
+                      let zipper = Zipper.move kind zipper in
+                      (zipper, res)
+                in
                 (zipper, res)
           in
           (zipper, res)
   )
 
-let read_comics mode (archives : Comic.named_archive list) () =
+let read_comics ~config mode (archives : Comic.named_archive list) () =
   let () = Termove.start_window () in
   let () = Termove.hide_cursor () in
   let () = MagickWand.magick_wand_genesis () in
 
+  (* let () = archives |> List.map (fun s -> s.Comic.archive_path) |> String.concat "\n" |> debug_string in *)
   let collection = List.map Either.right archives in
   let z_collections = Zipper.of_list collection in
 
-  let _side = read_collection mode z_collections in
+  let _side = read_collection mode config z_collections in
 
   let () = Termove.end_window () in
   let () = Termove.show_cursor () in