+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+# 0.2.0
+- [Decrypt/Rename collection subcommand](https://github.com/EruEri/oyomu/pull/3)
+- [Add option to read commands for unzip keep strategy](https://github.com/EruEri/oyomu/pull/2)
+- [Use environment variable to change keybinding ](https://github.com/EruEri/oyomu/pull/1)
 
 # 0.1.0
 - Inilial release